changed s_a biasing weight from float to double

+++ src/calc_n_a/clone.h
@@ -15,7 +15,7 @@ typedef struct Clone_s
 
 	float wDotIntegral = 0.0 ;
 
-	float s_a ;
+	double s_a ;
 
 	int n_a ;
